+++ src/assets/matchSetup/CodeSubmission.csv
@@ -1,5 +1,4 @@
 ﻿ID,Start time,Completion time,Email,Name,Team Name,School,Home Monster Choice,Away Monster Choice,URL/Code
-1,9/29/20 14:50:15,9/29/20 14:50:48,anonymous,,Jerrod's Team,Colorado School of Mines,grouchy,triclops,This is my code
-1,9/29/20 14:50:15,9/29/20 14:50:48,anonymous,,Buff Coders,University of Colorado,bobo,goldy,This is my code
-1,9/29/20 14:50:15,9/29/20 14:50:48,anonymous,,Ram Coders,Colorado State University,goldy,bobo,This is my code
-1,9/29/20 14:50:15,9/29/20 14:50:48,anonymous,,Team Name,Metro,pinky,spike,This is my code
+1,9/29/20 14:50:15,9/29/20 14:50:48,anonymous,,Jerrods Team,Colorado School of Mines,grouchy,triclops,https://raw.githubusercontent.com/SarahPerson/DropperAIs/master/wander.ai.ts
+1,9/29/20 14:50:15,9/29/20 14:50:48,anonymous,,Tile Status Team,Tyler Technologies,bobo,goldy,https://raw.githubusercontent.com/SarahPerson/DropperAIs/master/tile-status.ai.ts
+1,9/29/20 14:50:15,9/29/20 14:50:48,anonymous,,Invalid Code,Some School,goldy,bobo,This is my code
